bias-loan/requirements.txt
# local package
-e .

# external requirements
click==8.1.3
Sphinx
coverage==7.2.1
awscli==1.27.83
flake8==6.0.0
python-dotenv>=0.5.1
pandas==1.5.3